+++ Methods/DatabaseReader.js
@@ -5,10 +5,10 @@ class DatabaseReader {
     this.location = "./"+(typeof location === "string" ? location.trim() : "/ECMAData")
     if (create) {
       this.name = typeof location === "string" ? name.trim() : "ECMAData"
-      this.createDatabaseFolder()
+      this.#createDatabaseFolder()
     }
   }
-  createDatabaseFolder() {
+  #createDatabaseFolder() {
     try {
       let config
       if (fs.existsSync(this.location)) {
@@ -103,7 +103,6 @@ class DatabaseReader {
         fs.writeFileSync(`${this.location}/${tableName}.json`, JSON.stringify([]))
       }
     }
-    else console.error("Table already created")
   }
   removeTable(tableName) {
     delete this.config.tables[tableName]
@@ -114,8 +113,86 @@ class DatabaseReader {
       console.error(`Cannot remove table "${tableName}"`)
     }
   }
-  insertColumn(tableName, column) {
-
+  insertColumn(tableName, columns) {
+    let tables = this.getConfig().tables
+    let breaking = false
+    if (tables[tableName]) {
+      let tableColumns = {}
+      columns.forEach((columnName, properties) => {
+        if (!properties.type) {
+          console.error(`No type specified for "${columnName}"`)
+          breaking = true
+          return
+        }
+        let Spec = this.getPropSpecs(properties.type)
+        if (Spec) {
+          properties.forEach((attribute, attrVal) => {
+            if (!Spec[attribute](attrVal)) {
+              console.error(`Wrong attributes for property "${attribute}"`)
+              breaking = true
+              return
+            }
+          })
+          if (!breaking) {
+            tableColumns = {
+              ...tableColumns,
+              [columnName]: properties
+            }
+          }
+        }
+        else breaking = true
+      });
+      if (!breaking) {
+        tables = {
+          ...tables,
+          [tableName]: {
+            ...tables[tableName], 
+            ...tableColumns
+          }
+        }
+        this.config.tables = tables;
+        fs.writeFileSync(`${this.location}/config.json`, JSON.stringify(this.getConfig()))
+      }
+    }
+    else console.error("Table didn't exist")
+  }
+  updateColumn(tableName, columnName, columnProps) {
+    let tableConfig = this.getConfig().tables[tableName][columnName]
+    let prop = {}, isValid = true
+    let Spec = this.getPropSpecs(tableConfig.type)
+    columnProps.forEach((propName, propValue) => {
+      if (Spec[propName](propValue))
+        prop = { ...prop, [propName]: propValue }
+      else { 
+        isValid = false
+        console.error(`Property ${propName} is not valid`)
+      }
+    })
+    if (isValid) {
+      this.config.tables[tableName][columnName] = {
+        ...tableConfig,
+        ...prop
+      }
+      fs.writeFileSync(`${this.location}/config.json`, JSON.stringify(this.getConfig()))
+    }
+  }
+  removeColumn(tableName, arrayColumn) {
+    let tablesConfig = this.getConfig().tables
+    if (typeof arrayColumn === "string") arrayColumn = [arrayColumn]
+    let table = JSON.parse(fs.readFileSync(`${this.location}/${tableName}.json`, {encoding: 'utf-8'}))
+    if (tablesConfig[tableName]) {
+      arrayColumn.forEach(columnName => {
+        table.forEach((t, index) => {
+          delete table[index][columnName]
+        })
+        fs.writeFileSync(`${this.location}/${tableName}.json`, JSON.stringify(table))
+        if (!this.getConfig().tables[tableName][columnName])
+          console.error(`Column "${columnName}" didn't exist`)
+        delete this.getConfig().tables[tableName][columnName]
+      })
+      fs.writeFileSync(`${this.location}/config.json`, JSON.stringify(this.getConfig()))
+    }
+    else console.error("Table didn't exist")
   }
 }